requirements.txt
moviepy==1.0.3
openai==1.52.2
openai-whisper==20240930
